requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
clickhouse-connect==0.7.0
coloredlogs==15.0.1
comm==0.2.1
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
duckdb==0.9.2
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.27.0
googleapis-common-protos==1.62.0
grandalf==0.8
graphlib-backport==1.0.3
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hnswlib==0.8.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter-client==8.6.0
jupyter-core==5.7.1
kubernetes==29.0.0
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-openai==0.0.5
langchainhub==0.1.14
langserve==0.0.41
langsmith==0.0.86
lark==1.1.9
lz4==4.3.3
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.11.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.13
overrides==7.7.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pddl==0.4.0
PDDL-Prover==1.0.0
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
portalocker==2.8.2
posthog==3.4.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic-core==2.16.2
pygments==2.17.2
pyparsing==3.1.1
PyPika==0.48.9
pyproject-hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.7.2
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tavily-python==0.3.1
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.1
tomli==2.0.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.2.0
typer==0.9.0
types-requests==2.31.0.20240125
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
seaborn